Add files via upload
Firstly, we analyze the factors/ criteria banks use in assessing individuals as fit for loans. I analyzed the distribution of each feature and then explored the account holders’ distribution. 
Then I analyzed Income distribution and whether education impacts availing loans or not. Lastly, I performed hypothesis testing to check whether the number of persons in the individual’s family has a significant impact on availing of personal loans or not.